requirements.txt
customtkinter>=5.2.0
webdavclient3>=3.14.6
requests>=2.31.0
packaging>=23.0
pyinstaller>=6.0.0
oletools>=0.60.0
pyyaml>=6.0